Fix verified static-analysis bugs; reject false positives
Independently audited ~111 AI-reported bugs against origin/master and the
two fix branches (ai_bugfixes, copilot/fix-bugs-from-bugs-md). Applied the
correct fix for each genuine defect; rejected non-bugs and regressive/
incorrect fixes. Full per-bug audit in BUG_VERIFICATION.md.

Rejected as NOT-A-BUG: 40 (FFT flag convention already correct; both
branches' change is a regression), 84 (speculative int widening; total
size already bigint), 93 (int*sizeof already promotes to 64-bit), 102
(bounded buffer, no overflow).

Corrected branch mistakes: avoided ai_bugfixes' deletion of the EPSZERO
divide-by-zero guard in collide_vss.cpp / collide_vss_kokkos.cpp (bug 46);
used the correct scratch-variable fix for react_qk rejection sampling
(bug 72) rather than copilot's react_prob-polluting version.

Completeness: ported the verified fixes for bugs 62/77/78/79/80 to the CPU
copies (update.cpp, geometry.cpp) that both branches fixed only in KOKKOS;
completed partial fixes (57 double memset, 58 set2, 60 dbuf, 73 ambipolar
loop, 100 KOKKOS); applied 82/83 which neither branch fixed.

Build: make serial links spa_serial cleanly with all non-KOKKOS fixes.

+++ src/marching_cubes.h
@@ -33,7 +33,7 @@ class MarchingCubes : protected Pointers {
   double thresh;
 
   double *lo,*hi;
-  int v000,v001,v010,v011,v100,v101,v110,v111;
+  double v000,v001,v010,v011,v100,v101,v110,v111;
   double v000iso,v001iso,v010iso,v011iso,v100iso,v101iso,v110iso,v111iso;
   double inval[8][6];
   double i0,i1,i2,i3,i4,i5,i6,i7,i8,i9,i10,i11;
